Add initial scrypt KDF/password support.
This adds support for using scrypt for key stretching, and for securely
encrypting and validating passwords to disk (for example, in a web
application) with a high-level interface.

A lot of the code came from Falko Peters, who wrote all the initial code
for his 'scrypt' package. I merely appropriated it and shuffled some
stuff around for my own use. A lot of thanks to him!

The underlying primitive is not Colin Percival's scrypt, but Andrew M's
scrypt-jane. It has additional optimizations and enables the choice of
different mix/hash functions. We use scrypt-BLAKE512-ChaCha20/8 here (as
opposed to scrypt-SHA512-Salsa20/8.)

Currently the password support isn't exposed from the top-level module,
but the key stretching support is.

File encryption based on Colin Percival's `scryptenc` will come soon
(this will make it possible to use a password to encrypt a secret key on
disk for example.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>