chacha20: use chacha20-krovetz
So, my implementation was fast, but not fast enough. The SIMD
implementation by Ted Krovetz is, as far as I'm aware, both the fastest
and easiest implementation to use. This doubled the speed of my
networking protocol code. Combined with the previous commit, this
cranked up the speed from 20Mb/s -> 110Mb/s

Signed-off-by: Austin Seipp <aseipp@pobox.com>